requirements.txt
cupy-cuda12x 
numpy
pytest  
matplotlib 
tqdm  